Move notice component into summary panel
