+++ packages/manager/src/features/PlacementGroups/PlacementGroupsDetail/PlacementGroupsDetail.tsx
@@ -1,20 +1,16 @@
 import { AFFINITY_TYPES } from '@linode/api-v4';
-import { useTheme } from '@mui/material';
 import * as React from 'react';
 import { useHistory, useParams } from 'react-router-dom';
-import { Link } from 'react-router-dom';
 
 import { CircleProgress } from 'src/components/CircleProgress';
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 import { ErrorState } from 'src/components/ErrorState/ErrorState';
 import { LandingHeader } from 'src/components/LandingHeader';
 import { NotFound } from 'src/components/NotFound';
-import { Notice } from 'src/components/Notice/Notice';
 import { SafeTabPanel } from 'src/components/Tabs/SafeTabPanel';
 import { TabLinkList } from 'src/components/Tabs/TabLinkList';
 import { TabPanels } from 'src/components/Tabs/TabPanels';
 import { Tabs } from 'src/components/Tabs/Tabs';
-import { Typography } from 'src/components/Typography';
 import { useFlags } from 'src/hooks/useFlags';
 import {
   useMutatePlacementGroup,
@@ -23,7 +19,6 @@ import {
 import { getErrorStringOrDefault } from 'src/utilities/errorUtils';
 
 import { getPlacementGroupLinodeCount } from '../utils';
-import { getWarningNoticeText } from '../utils';
 import { PlacementGroupsLinodes } from './PlacementGroupsLinodes/PlacementGroupsLinodes';
 import { PlacementGroupsSummary } from './PlacementGroupsSummary/PlacementGroupsSummary';
 
@@ -31,7 +26,6 @@ export const PlacementGroupsDetail = () => {
   const flags = useFlags();
   const { id, tab } = useParams<{ id: string; tab?: string }>();
   const history = useHistory();
-  const theme = useTheme();
   const placementGroupId = Number(id);
 
   const {
@@ -119,21 +113,6 @@ export const PlacementGroupsDetail = () => {
         <TabLinkList tabs={tabs} />
         <TabPanels>
           <SafeTabPanel index={0}>
-            {!placementGroup.compliant && (
-              <Notice spacingBottom={20} spacingTop={24} variant="warning">
-                <Typography fontFamily={theme.font.bold}>
-                  {getWarningNoticeText(placementGroup)}
-                  {/* TODO VM_Placement: Get link location */}
-                  <Link
-                    className="secondaryLink"
-                    data-testid="pg-non-compliant-notice-link"
-                    to={'#'}
-                  >
-                    Learn more.
-                  </Link>
-                </Typography>
-              </Notice>
-            )}
             <PlacementGroupsSummary placementGroup={placementGroup} />
           </SafeTabPanel>
           <SafeTabPanel index={1}>
